requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
Django==1.9.1
django-allauth==0.24.1
django-bootstrap3==6.2.2
django-braces==1.8.1
django-ckeditor==5.0.2
django-model-utils==2.4
django-notifications-hq==1.0
django-postgrespool==0.3.0
gunicorn==19.4.5
jsonfield==1.0.3
oauthlib==1.0.3
psycopg2==2.6.1
python3-openid==3.0.9
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
SQLAlchemy==1.0.11
static3==0.6.1
wheel==0.24.0
whitenoise==2.0.6
